Reports: split component list in Template Builder into two tabs
